Split address bar into its own component. Un-hardcode css widths. Star!
